chocanon::db

Struct DB

pub struct DB { /* private fields */ }

A ChocAn database.

Implementations

impl DB

pub fn new(path: &str) -> Result<Self, Error>

Create a ChocAn database

Arguments

path - The path to the database file (E.g. ./database.db3).

Failure

Will return Err if database could not be established.

pub fn send_member_reports(&self) -> Result<(), Error>

Sends out all member reports to all ChocAn members.

Reports will only be sent to those with recent activity. Reports will only be sent valid persons.

Failure

Will return Err if any reports are not sent.

pub fn send_provider_reports(&self) -> Result<(), Error>

Sends out all provider reports to all ChocAn providers.

Reports will only be sent to those with recent activity. Reports will only be sent valid persons.

Failure

Will return Err if any reports are not sent.

pub fn send_manager_report(&self) -> Result<(), Error>

Sends out a manager report to the ChocAn manager.

Failure

Will return Err if any reports are not sent.

pub fn send_provider_directory(&self, email: &str) -> Result<(), Error>

Sends out a the provider directory to the specified email.

Arguments

email - The email address to send the provider directory to.

Failure

Will return Err if not sent.

pub fn is_valid_member_id(&self, id: u32) -> Result<bool, Error>

Checks if the member id belongs to a member in the database.

Arguments

id - The member id to check.

Success

Will return Ok wrapping true if valid person found. Otherwise, will return Ok wrapping false.

Failure

Will return Err on database error.

pub fn is_valid_provider_id(&self, id: u32) -> Result<bool, Error>

Checks if the provider id belongs to a provider in the database.

Arguments

id - The provider id to check.

Success

Will return Ok wrapping true if valid person found. Otherwise, will return Ok wrapping false.

Failure

Will return Err on database error.

pub fn is_valid_service_id(&self, id: u32) -> Result<bool, Error>

Checks if the service id belongs to a service in the directory.

Arguments

id - The service id to check.

Success

Will return Ok wrapping true if valid service found. Otherwise, will return Ok wrapping false.

Failure

Will return Err on database error.

pub fn add_member(&self, person: &PersonInfo) -> Result<(), Error>

Adds a member to the database.

Arguments

person - The member to add to the database.

Failure

Will return Err if the member was not added.

pub fn add_provider(&self, person: &PersonInfo) -> Result<(), Error>

Adds a provider to the database.

Arguments

person - The provider to add to the database.

Failure

Will return Err if the provider was not added.

pub fn remove_member(&self, id: u32) -> Result<(), Error>

Removes a member from the database.

Arguments

id - The id of the member to remove.

Failure

Will return Err if the member was not removed.

pub fn remove_provider(&self, id: u32) -> Result<(), Error>

Removes a provider from the database.

Arguments

id - The id of the provider to remove.

Failure

Will return Err if the provider was not removed.

pub fn add_consultation_record( &self, consul: &Consultation, ) -> Result<(), Error>

Adds a consultation record to the database.

Arguments

consul - The consultation to add to the database.

Failure

Will return Err if the record was not added.

pub fn add_service(&self, id: u32, name: &str, fee: f64) -> Result<(), Error>

Adds a service to the provider directory.

Arguments

id - The id of the service.

name - The name of the service.

fee - The fee for the service.

Failure

Will return Err if the service was not added.

pub fn get_service_name(&self, id: u32) -> Result<String, Error>

Gets the name corresponding to the specified service code id.

Arguments

id - The id of the service to get the name of.

Failure

Will return Err if the name could not be retrieved.

pub fn get_service_fee(&self, id: u32) -> Result<f64, Error>

Gets the fee corresponding to the specified service code id.

Arguments

id - The id of the service to get the fee of.

Failure

Will return Err if the fee could not be retrieved.

pub fn get_member_info(&self, id: u32) -> Result<PersonInfo, Error>

Gets the data for a member in the database.

Arguments

id - The id of the member to get the data for.

Failure

Will return Err if the data could not be retrieved.

pub fn get_provider_info(&self, id: u32) -> Result<PersonInfo, Error>

Gets the data for a provider in the database.

Arguments

id - The id of the provider to get the data for.

Failure

Will return Err if the data could not be retrieved.